refactor: Removed tf typing
