+++ doctr/datasets/ocr.py
@@ -8,9 +8,8 @@ import json
 import numpy as np
 from pathlib import Path
 from typing import List, Dict, Any, Tuple, Optional, Callable
-import tensorflow as tf
 
-from .core import AbstractDataset
+from .datasets import AbstractDataset
 from doctr.utils.geometry import fit_rbbox
 
 
@@ -32,7 +31,7 @@ class OCRDataset(AbstractDataset):
         self,
         img_folder: str,
         label_file: str,
-        sample_transforms: Optional[Callable[[tf.Tensor], tf.Tensor]] = None,
+        sample_transforms: Optional[Callable[[Any], Any]] = None,
         rotated_bbox: bool = False,
         **kwargs: Any,
     ) -> None:
